+++ packages/charts/react-charts/library/src/components/DeclarativeChart/PlotlySchemaAdapter.ts
@@ -1,3 +1,5 @@
+'use client';
+
 /* eslint-disable one-var */
 /* eslint-disable vars-on-top */
 /* eslint-disable no-var */