+++ packages/react-components/react-progress/library/src/components/ProgressBar/useProgressBar.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import * as React from 'react';
 import { useFieldContext_unstable } from '@fluentui/react-field';
 import { getIntrinsicElementProps, slot } from '@fluentui/react-utilities';