+++ packages/react-components/react-tree/library/src/components/FlatTree/useFlatControllableCheckedItems.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useControllableState } from '@fluentui/react-utilities';
 import { TreeItemValue } from '../../TreeItem';
 import { ImmutableMap } from '../../utils/ImmutableMap';