+++ packages/react-components/react-tree/library/src/components/Tree/Tree.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import * as React from 'react';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
 import { useTree_unstable } from './useTree';