+++ packages/react-components/react-icons-compat/library/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - @fluentui/react-icons-compat
 
-This log was last generated on Mon, 08 Sep 2025 12:50:58 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Oct 2025 15:12:38 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.2.9](https://github.com/microsoft/fluentui/tree/@fluentui/react-icons-compat_v0.2.9)
+
+Thu, 02 Oct 2025 15:12:38 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-icons-compat_v0.2.8..@fluentui/react-icons-compat_v0.2.9)
+
+### Patches
+
+- Bump @fluentui/react-jsx-runtime to v9.2.1 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
+- Bump @fluentui/react-shared-contexts to v9.25.2 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
+- Bump @fluentui/react-utilities to v9.25.0 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
 
 ## [0.2.8](https://github.com/microsoft/fluentui/tree/@fluentui/react-icons-compat_v0.2.8)