+++ packages/react-components/react-provider/library/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @fluentui/react-provider
 
-This log was last generated on Mon, 08 Sep 2025 12:51:18 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Oct 2025 15:12:31 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.22.6](https://github.com/microsoft/fluentui/tree/@fluentui/react-provider_v9.22.6)
+
+Thu, 02 Oct 2025 15:12:31 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-provider_v9.22.5..@fluentui/react-provider_v9.22.6)
+
+### Patches
+
+- feat: enforce usage of use client directive for files with client-only features ([PR #35173](https://github.com/microsoft/fluentui/pull/35173) by dmytrokirpa@microsoft.com)
+- Bump @fluentui/react-shared-contexts to v9.25.2 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
+- Bump @fluentui/react-tabster to v9.26.6 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
+- Bump @fluentui/react-utilities to v9.25.0 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.2.1 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
 
 ## [9.22.5](https://github.com/microsoft/fluentui/tree/@fluentui/react-provider_v9.22.5)