+++ packages/react-components/react-timepicker-compat/library/CHANGELOG.md
@@ -1,8 +1,22 @@
 # Change Log - @fluentui/react-timepicker-compat
 
-This log was last generated on Mon, 08 Sep 2025 12:50:48 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Oct 2025 15:12:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.4.17](https://github.com/microsoft/fluentui/tree/@fluentui/react-timepicker-compat_v0.4.17)
+
+Thu, 02 Oct 2025 15:12:07 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-timepicker-compat_v0.4.16..@fluentui/react-timepicker-compat_v0.4.17)
+
+### Patches
+
+- feat: enforce usage of use client directive for files with client-only features ([PR #35173](https://github.com/microsoft/fluentui/pull/35173) by dmytrokirpa@microsoft.com)
+- Bump @fluentui/react-combobox to v9.16.7 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
+- Bump @fluentui/react-field to v9.4.6 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.2.1 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
+- Bump @fluentui/react-shared-contexts to v9.25.2 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
+- Bump @fluentui/react-utilities to v9.25.0 ([PR #35133](https://github.com/microsoft/fluentui/pull/35133) by beachball)
 
 ## [0.4.16](https://github.com/microsoft/fluentui/tree/@fluentui/react-timepicker-compat_v0.4.16)